Move diis.py
