+++ psi4/driver/procrouting/__init__.py
@@ -30,6 +30,7 @@ from .proc_table import procedures, hooks, energy_only_methods, integrated_basis
 from .proc import scf_helper, scf_wavefunction_factory
 from .empirical_dispersion import EmpiricalDispersion
 from . import dft
+from . import diis
+from . import libcubeprop
 from . import response
 from . import scf_proc
-from . import libcubeprop
