+++ psi4/driver/procrouting/scf_proc/__init__.py
@@ -30,4 +30,4 @@
 A helper folder for auxiliary SCF functions and iterations.
 """
 
-from . import scf_iterator, subclass_methods, diis
+from . import scf_iterator, subclass_methods
